# docs/api-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Singra API",
    "version": "1.0.0",
    "description": "API do Singra. Só JSON: quem desenha as telas é o frontend estático, publicado à parte. Toda rota que não seja de autenticação espera o cabeçalho Authorization: Bearer <token>, e devolve apenas os dados do dono do token."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://singra-api.onrender.com",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "tokenSingra": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token assinado devolvido pelo login. Vale 30 dias."
      }
    }
  },
  "paths": {
...
    "/api/categorias": {
      "get": {
        "summary": "Lista as categorias do usuário e as mais usadas.",
        "tags": [
          "geral"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "security": [
          {
            "tokenSingra": []
          }
        ]
      },
      "post": {
        "summary": "Cria uma categoria.",
        "tags": [
          "geral"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "security": [
          {
            "tokenSingra": []
          }
        ]
      }
    },
...
    "/api/painel": {
      "get": {
        "summary": "Tudo o que a tela Início mostra, em uma requisição só.",
        "tags": [
          "geral"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "security": [
          {
            "tokenSingra": []
          }
        ]
      }
    },
...
    "/api/resumo": {
      "get": {
        "summary": "Ganhos, gastos, disponível, taxa de poupança e guardado.",
        "tags": [
          "geral"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "security": [
          {
            "tokenSingra": []
          }
        ]
      }
    },
    "/api/transacoes": {
      "post": {
        "summary": "Registra um gasto ou um ganho, parcelando se pedido.",
        "tags": [
          "geral"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "security": [
          {
            "tokenSingra": []
          }
        ]
      },
      "get": {
        "summary": "Histórico, com os filtros da tela.",
        "tags": [
          "geral"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "security": [
          {
            "tokenSingra": []
          }
        ]
      }
    },
    "/api/transacoes/{transacao_id}": {
      "put": {
        "summary": "Corrige um registro.",
        "tags": [
          "transacoes"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "parameters": [
          {
            "name": "transacao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "tokenSingra": []
          }
        ]
      },
      "delete": {
        "summary": "Apaga um registro ou a compra parcelada inteira.",
        "tags": [
          "transacoes"
        ],
        "responses": {
          "200": {
            "description": "Deu certo."
          },
          "400": {
            "description": "Falta algum campo ou ele veio errado."
          },
          "401": {
            "description": "Sem token, ou com o token vencido."
          }
        },
        "parameters": [
          {
            "name": "transacao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "tokenSingra": []
          }
        ]
      }
    }
  }
}